Add NCrunch solution/project settings config files
NCrunch is a continuous unit test runner which provides developers with
a tight feedback loop by running impacted unit tests upon detecting
changes to the underlying source code.

NCrunch recommends that these files be checked into source code
control--they configure the projects and solution to work with NCrunch.
Normally, getting NCrunch to work with a solution is easy, just enable
it. But for Git-Tfs, it was a bit more complicated because of the
architecture and because of LibGit2Sharp's unmanaged dependency. So,
checking in these files will ensure that any other developers using
NCrunch will be all set to go.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>